refactor rpki module
